Fixed import error and added MAC getch()
